clean code scene 1 part 1
